Progress: add groups
So far: the add groups functions works like it should. Next: show the
form that let’s users choose if we should add ArcGIS Users to groups
and if so, show the form that asks which field contains those fields.
When the form is submitted, all existing users should be added and all
non-existing users should be created and added